Enable indexer Acknowledgement in the PubSub CF
To make sure the events sent to Splunk were indexed,
this patch add functionality to clear the status of an event after
sending to HEC using the indexer Acknowledgement feature.
if events did not make it through the index, the CF will attempt to
send it to the retry topic to avoid data loss.